+++ streamflow/deployment/connector/ssh.py
@@ -538,7 +538,7 @@ class SSHConnector(BaseConnector):
                                 self.hardware[location].storage[mount_point] = Storage(
                                     mount_point=mount_point,
                                     size=float(size) / 2**10,
-                                    inmemory_usage=(
+                                    memory_usage=(
                                         0.0 if fs_type in IN_MEMORY_FS_TYPES else None
                                     ),
                                 )